Rename constants.ts => mock.ts
